pyproject.toml
[project]
name = "imgmodaldb"
version = "0.1.0"
description = "Desktop application for multimodal image search with pluggable embedders and vector stores."
readme = "Dev.md"
authors = [{ name = "ImgModalDB Team" }]
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26",
    "pillow>=10.0",
    "pydantic>=2.6",
    "tqdm>=4.66",
    "loguru>=0.7",
]

[project.optional-dependencies]
embedders = [
    "open-clip-torch>=2.24.0",
    "jina-embeddings-v4>=0.0.26",
]
vector_stores = [
    "faiss-cpu>=1.7.4",
]
gui = [
    "PySide6>=6.6",
]
api = [
    "fastapi>=0.111",
]
tests = [
    "pytest>=8.0",
]

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"
